feat(devtui): modal value picker and overlay restructure
Pressing Enter on a Config tab field now opens a centered modal
(list picker for selects, text input for credentials) instead of
cycling inline. ←/→ cycle is removed.

Internally, dismissable overlays (palette, picker, stop-confirm)
are unified behind a Modal interface separate from lifecycle phases,
and model_update.go is split into lifecycle.go, install_wizard.go,
task_runner.go, keys.go, and overlay_*.go so each file has one
purpose and adding a new tab/modal/phase touches one place.
