+++ internal/devtui/keys.go
@@ -1,0 +1,19 @@
+package devtui
+
+const (
+	keyCtrlC    = "ctrl+c"
+	keyDown     = "down"
+	keyEnter    = "enter"
+	keyUp       = "up"
+	keyTab      = "tab"
+	keyShiftTab = "shift+tab"
+	keyQ        = "q"
+	keyF        = "f"
+	keyJ        = "j"
+	keyK        = "k"
+	key1        = "1"
+	key2        = "2"
+	key3        = "3"
+	keyLeft     = "left"
+	keyRight    = "right"
+)
