+++ internal/devtui/lifecycle.go
@@ -1,0 +1,119 @@
+package devtui
+
+import (
+	"strings"
+
+	"charm.land/bubbles/v2/textinput"
+	tea "charm.land/bubbletea/v2"
+
+	"github.com/shopware/shopware-cli/internal/shop"
+)
+
+// updateLifecycle handles async messages from the boot state machine: Docker
+// container start/stop, Shopware install detection, install completion. The
+// state machine drives m.phase forward; user input during a phase is handled
+// by phase-specific key handlers in updateKeyPress.
+func (m Model) updateLifecycle(msg tea.Msg) (tea.Model, tea.Cmd) {
+	switch msg := msg.(type) {
+	case dockerAlreadyRunningMsg:
+		m.phase = phaseDashboard
+		return m, m.checkShopwareInstalled()
+
+	case dockerNeedStartMsg:
+		m.phase = phaseStarting
+		m.overlayLines = nil
+		m.dockerShowLogs = false
+		m.dockerSpinner = newBrandSpinner()
+		return m, tea.Batch(m.dockerSpinner.Tick, m.startContainers())
+
+	case dockerOutputLineMsg:
+		m.overlayLines = append(m.overlayLines, string(msg))
+		maxLines := m.overlayMaxLines()
+		if len(m.overlayLines) > maxLines {
+			m.overlayLines = m.overlayLines[len(m.overlayLines)-maxLines:]
+		}
+		if m.phase == phaseInstalling {
+			line := string(msg)
+			if strings.HasPrefix(line, "Start: ") {
+				for i, sp := range installStepPatterns {
+					if strings.Contains(line, sp.pattern) && i >= m.installProg.currentStep {
+						m.installProg.currentStep = i
+						pct := float64(i) / float64(len(installStepPatterns))
+						cmd := m.installProg.progress.SetPercent(pct)
+						return m, tea.Batch(cmd, m.readNextDockerOutput())
+					}
+				}
+			}
+		}
+		return m, m.readNextDockerOutput()
+
+	case dockerOutputDoneMsg:
+		return m, nil
+
+	case dockerStartedMsg:
+		if msg.err != nil {
+			m.overlayLines = append(m.overlayLines, errorStyle.Render("Failed: "+msg.err.Error()))
+			m.overlayLines = append(m.overlayLines, "", helpStyle.Render("Press q to exit"))
+			return m, nil
+		}
+		m.phase = phaseDashboard
+		m.overlayLines = nil
+		m.dockerOutChan = nil
+		return m, m.checkShopwareInstalled()
+
+	case shopwareInstalledMsg:
+		m.phase = phaseDashboard
+		return m, m.startDashboard()
+
+	case shopwareNotInstalledMsg:
+		m.phase = phaseInstallPrompt
+		m.overlayLines = nil
+
+		usernameInput := textinput.New()
+		usernameInput.Placeholder = defaultUsername
+		usernameInput.Prompt = "Username: "
+		usernameInput.CharLimit = 50
+
+		passwordInput := textinput.New()
+		passwordInput.Placeholder = "shopware"
+		passwordInput.Prompt = "Password: "
+		passwordInput.CharLimit = 50
+		passwordInput.EchoMode = textinput.EchoPassword
+
+		m.install = installWizard{step: installStepAsk, confirmYes: true, username: usernameInput, password: passwordInput}
+		return m, nil
+
+	case shopwareInstallDoneMsg:
+		if msg.err != nil {
+			m.installProg.showLogs = true
+			m.overlayLines = append(m.overlayLines, "", errorStyle.Render("Installation failed: "+msg.err.Error()))
+			m.overlayLines = append(m.overlayLines, "", helpStyle.Render("Press q to exit"))
+			return m, nil
+		}
+		m.installProg.done = true
+		m.installProg.currentStep = len(installStepPatterns)
+
+		username := m.install.username.Value()
+		password := m.install.password.Value()
+
+		adminApi := &shop.ConfigAdminApi{
+			Username: username,
+			Password: password,
+		}
+		m.envConfig.AdminApi = adminApi
+		_ = shop.WriteConfig(m.config, m.projectRoot)
+
+		m.general.username = username
+		m.general.password = password
+
+		m.phase = phaseDashboard
+		m.overlayLines = nil
+		m.dockerOutChan = nil
+		return m, m.startDashboard()
+
+	case dockerStoppedMsg:
+		return m, tea.Quit
+	}
+
+	return m, nil
+}
